+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+_In progress…_
+
+## [1.2.0] - 2018-08-20
 ### Added
 - Add a `log_node_console` option to log the output of console methods (`console.log`, `console.debug`, `console.table`, etc…) to the PHP logger
 
@@ -60,7 +63,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 First release
 
 
-[Unreleased]: https://github.com/nesk/rialto/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/nesk/rialto/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/nesk/rialto/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/nesk/rialto/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/nesk/rialto/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/nesk/rialto/compare/1.0.0...1.0.1